pyproject.toml
[project]
name = "procat"
dynamic = ["version"]
description = "Project Charging and Analytics Tool"
authors = [
    { name = "Diego Alonso Alvarez", email = "d.alonso-alvarez@imperial.ac.uk" },
    { name = "Imperial College London RSE Team", email = "ict-rse-team@imperial.ac.uk" },
]
requires-python = ">=3.13"
dependencies = [
    "django",
    "django-bootstrap5",
    "django-crispy-forms",
    "whitenoise",
    "crispy-bootstrap5",
    "django_tables2",
    "django-filter",
    "huey",
    "types-requests",
    "requests",
    "bokeh",
    "pandas",
    "mozilla-django-oidc",
    "django-admin-rangefilter"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "mypy",
    "pip-tools",
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "pytest-mock",
    "django-stubs[compatible-mypy]",
    "djlint",
    "pandas-stubs",
]
doc = [
    "mkdocs",
    "mkdocstrings",
    "mkdocstrings-python",
    "mkdocs-material",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-section-index",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
disallow_any_explicit = true
disallow_any_generics = true
warn_unreachable = true
warn_unused_ignores = true
disallow_untyped_defs = true
exclude = [".venv/", "docs/", "main/migrations/"]

[tool.django-stubs]
django_settings_module = "procat.settings"

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = ["django_tables2.*", "django_filters.*", "mozilla_django_oidc.*", "rangefilter.*"]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = ["huey", "huey.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-v -p no:warnings --cov=main --cov=procat --cov-branch --cov-report=xml --cov-report=html --doctest-modules --ignore=procat/__main__.py --ignore=docs/ --ignore=procat/settings/"
DJANGO_SETTINGS_MODULE = "procat.settings"
FAIL_INVALID_TEMPLATE_VARS = true

[tool.ruff]
exclude = ["main/migrations"]
target-version = "py312"

[tool.ruff.lint]
select = [
    "D",   # pydocstyle
    "E",   # pycodestyle
    "F",   # Pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "RUF", # ruff
]
pydocstyle.convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D104"] # Missing docstring in public package

[tool.djlint]
profile = "django"
preserve_blank_lines = true
indent = 2

[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = [
    "htmlcov",
] # Exclude the coverage report file from setuptools package finder
